backend/requirements.txt
# API REST
fastapi==0.115.11
uvicorn==0.34.0
python-multipart==0.0.20
pydantic==2.10.6

# PyTorch
torch==2.5.1
torchvision==0.20.1

# Procesamiento de imágenes
Pillow==10.1.0

# Utilidades
numpy>=2.0.1